Moved parser implementations into subdirectories
